test: Move userinitialization and OutgoingMessageStore to new util module
