+++ test/test_game_fullgames.py
@@ -11,32 +11,13 @@
 
 import unittest
 import re
-import os.path
 from typing import Pattern, Dict, List
 
 import sqlalchemy
 import sqlalchemy.orm
-import toml
 
 from qaqa_bot import model, game
-from qaqa_bot.util import session_scope
-
-
-CONFIG = toml.load(os.path.join(os.path.dirname(__file__), "test_config.toml"))
-
-
-class OutgoingMessageStore:
-    """ A Mock class to test the message sending of a GameServer """
-    def __init__(self):
-        self.messages = []
-
-    def send_message(self, messages: List[game.Message]) -> None:
-        self.messages.extend(messages)
-
-    def fetch_messages(self) -> List[game.Message]:
-        current_messages = self.messages
-        self.messages = []
-        return current_messages
+from .util import CONFIG, OutgoingMessageStore, create_sample_users
 
 
 class FullGameTests(unittest.TestCase):
@@ -44,20 +25,10 @@ class FullGameTests(unittest.TestCase):
         # Setup database schema
         engine = sqlalchemy.create_engine(CONFIG['database']['connection'], echo=True)
         model.Base.metadata.create_all(engine)
+        create_sample_users(engine)
 
         self.message_store = OutgoingMessageStore()
         self.game_server = game.GameServer(CONFIG, self.message_store.send_message, engine)
-
-        # Use
-        users = [model.User(api_id=1, chat_id=11, name="Michael"),
-                      model.User(api_id=2, chat_id=12, name="Jenny"),
-                      model.User(api_id=3, chat_id=13, name="Lukas"),
-                      model.User(api_id=4, chat_id=14, name="Jannik")]
-
-        Session = sqlalchemy.orm.sessionmaker(bind=engine)
-        with session_scope(Session) as session:
-            for user in users:
-                session.add(user)
 
     TEXT_SUBMIT_RESPONSE = r"🆗"
